+++ x/leverage/keeper/exchange_rate_test.go
@@ -1,0 +1,30 @@
+package keeper_test
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	umeeapp "github.com/umee-network/umee/v2/app"
+)
+
+func (s *IntegrationTestSuite) TestDeriveExchangeRate() {
+	// The init scenario is being used so module balance starts at 1000 umee
+	// and the uToken supply starts at 1000 due to supplier account
+	_, addr := s.initBorrowScenario()
+
+	// artificially increase total borrows (by affecting a single address)
+	err := s.tk.SetBorrow(s.ctx, addr, sdk.NewInt64Coin(umeeapp.BondDenom, 2000000000)) // 2000 umee
+	s.Require().NoError(err)
+
+	// artificially set reserves
+	err = s.tk.SetReserveAmount(s.ctx, sdk.NewInt64Coin(umeeapp.BondDenom, 300000000)) // 300 umee
+	s.Require().NoError(err)
+
+	// expected token:uToken exchange rate
+	//    = (total borrows + module balance - reserves) / utoken supply
+	//    = 2000 + 1000 - 300 / 1000
+	//    = 2.7
+
+	// get derived exchange rate
+	rate := s.app.LeverageKeeper.DeriveExchangeRate(s.ctx, umeeapp.BondDenom)
+	s.Require().Equal(sdk.MustNewDecFromStr("2.7"), rate)
+}
