+++ x/leverage/keeper/keeper_test.go
@@ -689,257 +689,6 @@ func (s *IntegrationTestSuite) TestRepay_Overpay() {
 	s.Require().Error(err)
 }
 
-func (s *IntegrationTestSuite) TestDeriveExchangeRate() {
-	// The init scenario is being used so module balance starts at 1000 umee
-	// and the uToken supply starts at 1000 due to supplier account
-	_, addr := s.initBorrowScenario()
-
-	// artificially increase total borrows (by affecting a single address)
-	err := s.tk.SetBorrow(s.ctx, addr, sdk.NewInt64Coin(umeeapp.BondDenom, 2000000000)) // 2000 umee
-	s.Require().NoError(err)
-
-	// artificially set reserves
-	err = s.tk.SetReserveAmount(s.ctx, sdk.NewInt64Coin(umeeapp.BondDenom, 300000000)) // 300 umee
-	s.Require().NoError(err)
-
-	// expected token:uToken exchange rate
-	//    = (total borrows + module balance - reserves) / utoken supply
-	//    = 2000 + 1000 - 300 / 1000
-	//    = 2.7
-
-	// get derived exchange rate
-	rate := s.app.LeverageKeeper.DeriveExchangeRate(s.ctx, umeeapp.BondDenom)
-	s.Require().Equal(sdk.MustNewDecFromStr("2.7"), rate)
-}
-
-func (s *IntegrationTestSuite) TestAccrueZeroInterest() {
-	// The "supplier" user from the init scenario is being used because it
-	// already has 1k u/umee for collateral.
-	addr, _ := s.initBorrowScenario()
-
-	// user borrows 40 umee
-	err := s.app.LeverageKeeper.Borrow(s.ctx, addr, sdk.NewInt64Coin(umeeapp.BondDenom, 40000000))
-	s.Require().NoError(err)
-
-	// verify user's loan amount (40 umee)
-	loanBalance := s.app.LeverageKeeper.GetBorrow(s.ctx, addr, umeeapp.BondDenom)
-	s.Require().Equal(loanBalance, sdk.NewInt64Coin(umeeapp.BondDenom, 40000000))
-
-	// Because no time has passed since genesis (due to test setup) this will not
-	// increase borrowed amount.
-	err = s.app.LeverageKeeper.AccrueAllInterest(s.ctx)
-	s.Require().NoError(err)
-
-	// verify user's loan amount (40 umee)
-	loanBalance = s.app.LeverageKeeper.GetBorrow(s.ctx, addr, umeeapp.BondDenom)
-	s.Require().Equal(loanBalance, sdk.NewInt64Coin(umeeapp.BondDenom, 40000000))
-
-	// borrow APY at utilization = 4%
-	// when kink utilization = 80%, and base/kink APY are 0.02 and 0.22
-	borrowAPY := s.app.LeverageKeeper.DeriveBorrowAPY(s.ctx, umeeapp.BondDenom)
-	s.Require().Equal(sdk.MustNewDecFromStr("0.03"), borrowAPY)
-
-	// supply APY when borrow APY is 3%
-	// and utilization is 4%, and reservefactor is 20%, and OracleRewardFactor is 1%
-	// 0.03 * 0.04 * (1 - 0.21) = 0.000948
-	supplyAPY := s.app.LeverageKeeper.DeriveSupplyAPY(s.ctx, umeeapp.BondDenom)
-	s.Require().NoError(err)
-	s.Require().Equal(sdk.MustNewDecFromStr("0.000948"), supplyAPY)
-}
-
-func (s *IntegrationTestSuite) TestDynamicInterest() {
-	// Init scenario is being used because the module account (lending pool)
-	// already has 1000 umee.
-	addr, _ := s.initBorrowScenario()
-
-	umeeToken := newToken("uumee", "UMEE")
-	umeeToken.CollateralWeight = sdk.MustNewDecFromStr("1.0")     // to allow high utilization
-	umeeToken.LiquidationThreshold = sdk.MustNewDecFromStr("1.0") // to allow high utilization
-
-	s.Require().NoError(s.app.LeverageKeeper.SetTokenSettings(s.ctx, umeeToken))
-
-	// Base interest rate (0% utilization)
-	rate := s.app.LeverageKeeper.DeriveBorrowAPY(s.ctx, umeeapp.BondDenom)
-	s.Require().Equal(rate, sdk.MustNewDecFromStr("0.02"))
-
-	// user borrows 200 umee, utilization 200/1000
-	err := s.app.LeverageKeeper.Borrow(s.ctx, addr, sdk.NewInt64Coin(umeeapp.BondDenom, 200000000))
-	s.Require().NoError(err)
-
-	// Between base interest and kink (20% utilization)
-	rate = s.app.LeverageKeeper.DeriveBorrowAPY(s.ctx, umeeapp.BondDenom)
-	s.Require().Equal(rate, sdk.MustNewDecFromStr("0.07"))
-
-	// user borrows 600 more umee, utilization 800/1000
-	err = s.app.LeverageKeeper.Borrow(s.ctx, addr, sdk.NewInt64Coin(umeeapp.BondDenom, 600000000))
-	s.Require().NoError(err)
-
-	// Kink interest rate (80% utilization)
-	rate = s.app.LeverageKeeper.DeriveBorrowAPY(s.ctx, umeeapp.BondDenom)
-	s.Require().NoError(err)
-	s.Require().Equal(rate, sdk.MustNewDecFromStr("0.22"))
-
-	// user borrows 100 more umee, utilization 900/1000
-	err = s.app.LeverageKeeper.Borrow(s.ctx, addr, sdk.NewInt64Coin(umeeapp.BondDenom, 100000000))
-	s.Require().NoError(err)
-
-	// Between kink interest and max (90% utilization)
-	rate = s.app.LeverageKeeper.DeriveBorrowAPY(s.ctx, umeeapp.BondDenom)
-	s.Require().NoError(err)
-	s.Require().Equal(rate, sdk.MustNewDecFromStr("0.87"))
-
-	// user borrows 100 more umee, utilization 1000/1000
-	err = s.app.LeverageKeeper.Borrow(s.ctx, addr, sdk.NewInt64Coin(umeeapp.BondDenom, 100000000))
-	s.Require().NoError(err)
-
-	// Max interest rate (100% utilization)
-	rate = s.app.LeverageKeeper.DeriveBorrowAPY(s.ctx, umeeapp.BondDenom)
-	s.Require().NoError(err)
-	s.Require().Equal(rate, sdk.MustNewDecFromStr("1.52"))
-}
-
-func (s *IntegrationTestSuite) TestDynamicInterest_InvalidAsset() {
-	rate := s.app.LeverageKeeper.DeriveBorrowAPY(s.ctx, "uabc")
-	s.Require().Equal(rate, sdk.ZeroDec())
-}
-
-func (s *IntegrationTestSuite) TestGetEligibleLiquidationTargets_OneAddrOneAsset() {
-	// The "supplier" user from the init scenario is being used because it
-	// already has 1k u/umee enabled as collateral.
-	addr, _ := s.initBorrowScenario()
-
-	// user borrows 100 umee (max current allowed) user amount enabled as collateral * CollateralWeight
-	// = 1000 * 0.1
-	// = 100
-	err := s.app.LeverageKeeper.Borrow(s.ctx, addr, sdk.NewInt64Coin(umeeapp.BondDenom, 100000000))
-	s.Require().NoError(err)
-
-	zeroAddresses, err := s.app.LeverageKeeper.GetEligibleLiquidationTargets(s.ctx)
-	s.Require().NoError(err)
-	s.Require().Equal([]sdk.AccAddress{}, zeroAddresses)
-
-	// Note: Setting umee liquidation threshold to 0.05 to make the user eligible to liquidation
-	umeeToken := newToken("uumee", "UMEE")
-	umeeToken.CollateralWeight = sdk.MustNewDecFromStr("0.05")
-	umeeToken.LiquidationThreshold = sdk.MustNewDecFromStr("0.05")
-
-	s.Require().NoError(s.app.LeverageKeeper.SetTokenSettings(s.ctx, umeeToken))
-
-	targetAddress, err := s.app.LeverageKeeper.GetEligibleLiquidationTargets(s.ctx)
-	s.Require().NoError(err)
-	s.Require().Equal([]sdk.AccAddress{addr}, targetAddress)
-
-	// if it tries to borrow any other asset it should return an error
-	err = s.app.LeverageKeeper.Borrow(s.ctx, addr, sdk.NewInt64Coin(atomDenom, 1))
-	s.Require().Error(err)
-}
-
-func (s *IntegrationTestSuite) TestGetEligibleLiquidationTargets_OneAddrTwoAsset() {
-	// The "supplier" user from the init scenario is being used because it
-	// already has 1k u/umee enabled as collateral.
-	addr, _ := s.initBorrowScenario()
-
-	// user borrows 100 umee (max current allowed) user amount enabled as collateral * CollateralWeight
-	// = 1000 * 0.1
-	// = 100
-	err := s.app.LeverageKeeper.Borrow(s.ctx, addr, sdk.NewInt64Coin(umeeapp.BondDenom, 100000000))
-	s.Require().NoError(err)
-
-	zeroAddresses, err := s.app.LeverageKeeper.GetEligibleLiquidationTargets(s.ctx)
-	s.Require().NoError(err)
-	s.Require().Equal([]sdk.AccAddress{}, zeroAddresses)
-
-	// mints and send to addr 100 atom and already
-	// enable 50 u/atom as collateral.
-	s.fundAccount(addr, coin(atomDenom, 100_000000))
-	s.supply(addr, coin(atomDenom, 50_000000))
-	s.collateralize(addr, coin("u/"+atomDenom, 50_000000))
-
-	// user borrows 4 atom (max current allowed - 1) user amount enabled as collateral * CollateralWeight
-	// = (50 * 0.1) - 1
-	// = 4
-	err = s.app.LeverageKeeper.Borrow(s.ctx, addr, sdk.NewInt64Coin(atomDenom, 4000000)) // 4 atom
-	s.Require().NoError(err)
-
-	// Note: Setting umee liquidation threshold to 0.05 to make the user eligible for liquidation
-	umeeToken := newToken("uumee", "UMEE")
-	umeeToken.CollateralWeight = sdk.MustNewDecFromStr("0.05")
-	umeeToken.LiquidationThreshold = sdk.MustNewDecFromStr("0.05")
-
-	s.Require().NoError(s.app.LeverageKeeper.SetTokenSettings(s.ctx, umeeToken))
-
-	// Note: Setting atom collateral weight to 0.01 to make the user eligible for liquidation
-	atomIBCToken := newToken(atomDenom, "ATOM")
-	atomIBCToken.CollateralWeight = sdk.MustNewDecFromStr("0.01")
-	atomIBCToken.LiquidationThreshold = sdk.MustNewDecFromStr("0.01")
-
-	s.Require().NoError(s.app.LeverageKeeper.SetTokenSettings(s.ctx, atomIBCToken))
-
-	targetAddr, err := s.app.LeverageKeeper.GetEligibleLiquidationTargets(s.ctx)
-	s.Require().NoError(err)
-	s.Require().Equal([]sdk.AccAddress{addr}, targetAddr)
-}
-
-func (s *IntegrationTestSuite) TestGetEligibleLiquidationTargets_TwoAddr() {
-	// The "supplier" user from the init scenario is being used because it
-	// already has 1k u/umee enabled as collateral.
-	supplierAddr, anotherSupplier := s.initBorrowScenario()
-
-	// supplier borrows 100 umee (max current allowed) supplier amount enabled as collateral * CollateralWeight
-	// = 1000 * 0.1
-	// = 100
-	err := s.app.LeverageKeeper.Borrow(s.ctx, supplierAddr, sdk.NewInt64Coin(umeeapp.BondDenom, 100000000))
-	s.Require().NoError(err)
-
-	zeroAddresses, err := s.app.LeverageKeeper.GetEligibleLiquidationTargets(s.ctx)
-	s.Require().NoError(err)
-	s.Require().Equal([]sdk.AccAddress{}, zeroAddresses)
-
-	// mints and send to anotherSupplier 100 atom and already
-	// enable 50 u/atom as collateral.
-	s.fundAccount(anotherSupplier, coin(atomDenom, 100_000000))
-	s.supply(anotherSupplier, coin(atomDenom, 50_000000))
-	s.collateralize(anotherSupplier, coin("u/"+atomDenom, 50_000000))
-
-	// anotherSupplier borrows 4 atom (max current allowed - 1) anotherSupplier amount enabled as collateral * CollateralWeight
-	// = (50 * 0.1) - 1
-	// = 4
-	err = s.app.LeverageKeeper.Borrow(s.ctx, anotherSupplier, sdk.NewInt64Coin(atomDenom, 4000000)) // 4 atom
-	s.Require().NoError(err)
-
-	// Note: Setting umee liquidation threshold to 0.05 to make the supplier eligible for liquidation
-	umeeToken := newToken("uumee", "UMEE")
-	umeeToken.CollateralWeight = sdk.MustNewDecFromStr("0.05")
-	umeeToken.LiquidationThreshold = sdk.MustNewDecFromStr("0.05")
-
-	s.Require().NoError(s.app.LeverageKeeper.SetTokenSettings(s.ctx, umeeToken))
-
-	// Note: Setting atom collateral weight to 0.01 to make the supplier eligible for liquidation
-	atomIBCToken := newToken(atomDenom, "ATOM")
-	atomIBCToken.CollateralWeight = sdk.MustNewDecFromStr("0.01")
-	atomIBCToken.LiquidationThreshold = sdk.MustNewDecFromStr("0.01")
-
-	s.Require().NoError(s.app.LeverageKeeper.SetTokenSettings(s.ctx, atomIBCToken))
-
-	supplierAddress, err := s.app.LeverageKeeper.GetEligibleLiquidationTargets(s.ctx)
-	s.Require().NoError(err)
-	s.Require().Equal([]sdk.AccAddress{supplierAddr, anotherSupplier}, supplierAddress)
-}
-
-func (s *IntegrationTestSuite) TestTotalCollateral() {
-	// Test zero collateral
-	uDenom := types.ToUTokenDenom(umeeDenom)
-	collateral := s.app.LeverageKeeper.GetTotalCollateral(s.ctx, uDenom)
-	s.Require().Equal(sdk.ZeroInt(), collateral)
-
-	// Uses borrow scenario, because supplier possesses collateral
-	_, _ = s.initBorrowScenario()
-
-	// Test nonzero collateral
-	collateral = s.app.LeverageKeeper.GetTotalCollateral(s.ctx, uDenom)
-	s.Require().Equal(sdk.NewInt(1000000000), collateral)
-}
-
 func (s *IntegrationTestSuite) TestLiqudate() {
 	addr, _ := s.initBorrowScenario()
